添加图层更新地图

+++ MyGis/Core/Semaphore.cpp
@@ -3,7 +3,7 @@
 namespace lh
 {
 
-Semaphore::Semaphore(int n)
+Semaphore::Semaphore(size_t n)
     : m_number(n)
 {
 }
@@ -12,14 +12,14 @@ Semaphore::~Semaphore()
 {
 }
 
-void Semaphore::release(int n)
+void Semaphore::release(size_t n)
 {
     std::lock_guard<std::mutex> lock(m_mutex);
     m_number += n;
     m_cv.notify_all();
 }
 
-void Semaphore::acquire(int n)
+void Semaphore::acquire(size_t n)
 {
     std::unique_lock<std::mutex> lock(m_mutex);
     while(m_number < n)
