+++ MyGis/Core/Semaphore.h
@@ -9,16 +9,16 @@ namespace lh
 class Semaphore
 {
 public:
-    explicit Semaphore(int n = 0);
+    explicit Semaphore(size_t n = 0);
     ~Semaphore();
 
-    void release(int n = 1);
-    void acquire(int n = 1);
+    void release(size_t n = 1);
+    void acquire(size_t n = 1);
 
 private:
     std::mutex m_mutex;
     std::condition_variable m_cv;
-    int m_number;
+    size_t m_number;
 };
 
 }
